requirements.txt
django~=5.0.2
gunicorn
psycopg2-binary
djangorestframework~=3.14.0
django-auth-adfs
python-dotenv
coverage
flake8
